requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.2
attrs==25.4.0
audioread==3.0.1
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.40.46
botocore==1.40.46
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-crontab==0.7.1
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.1
Flask==3.1.2
frozenlist==1.8.0
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
mutagen==1.47.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.23
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMuPDF==1.26.4
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.3
qrcode==8.2
regex==2025.11.3
requests==2.32.5
s3transfer==0.14.0
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
soupsieve==2.8.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
twilio==9.9.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
wikipedia==1.4.0
yarl==1.22.0
